Cleaned up LogHandler, improved performance, always include timestamp.
I chose to include both the date and the milliseconds. The date was mostly for use in a log file, whereas the milliseconds are needed when we want to diagnose performance issues of the API.
Performance of the log handler should also be increased by not re-opening the file writer every time we log something.
This change shouldn't cause issues with file handles remaining open - either the logger is closed explicitly, in which case the file handle is closed in the close() method, or the JVM is shut down forcefully, in which case all open file handles are released as well.