Early version of the reconnect feature.
This will probably be not be the final version, but should be is functional as of now.

TS3Config.setLoginCredentials had to be removed. Use TS3(Async)Api.login instead. I might add it in again at a later time to restore backwards-compatibility.

There has also been quite a bit of internal refactoring. If you've directly been accessing the IO part of the API, your code will break. If you absolutely need to have access to these parts of the API, let me know. Else the IO part will stay hidden from the user.