Fix a bug that would cause callbacks of reset commands to misbehave.
It turned out that creating a clean copy of a command would cause problems as the callback is expecting a different command object.
To avoid this problem we actually set all fields back to their initial values instead of creating a new command.
This solution is not quite as clean as the previous one, but it's simpler, has less side effects and actually works.