Remove unused DefaultListener
DefaultListener was never used, included bad code and sent messages to System.out instead of the logger. Some methods just did nothing. All events are already logged properly if the debug level is set low enough. There isn't an internal need for this class anymore.

If logging events over to System.out is absolutely needed, this should instead be implemented by the user and not the API.

For the reasons stated above, I decided to remove this class.