Add await and awaitUninterruptiby to CommandFuture
These methods let you wait for a CommandFuture to complete without calling get, which would throw an exception if the future is cancelled or the TS3 server returns an error.

Also added a short note to the javadoc that reminds users that all await and get methods are blocking and thus aren't asynchronous.