Add missing resources so examples can be run as-is
Previously, one could (usually) not run the examples as-is after having
cloned the repository. This was because there was no dependency on a
logging provider such as slf4j-simple when the "full" profile was
disabled. Moreover, the simplelogger config file in src/main/resources
could not be found when running the examples. This was fixed by simply
copying that file into the examples directory.