Ignore CRC32 icon ID collisions in uploadIcon and uploadIconDirect
If a query attempts to upload an icon that already exists on the server,
the query should ignore the "file already exists" error and return the
icon ID as if the upload had succeeded.

The chance of an actual CRC32 collision, where two different icons are
mapped to the same icon ID, is so tiny that we may as well ignore it.